-moved colorControls to popup/components/ -moved pages to popup/views
