+++ src/components/header/Header.tsx
@@ -9,7 +9,7 @@ export function Header({ page, setPage }: HeaderProps): JSX.Element {
     return (
         <div className="p-3 text-center relative">
             <h1 className="text-lg">ChatGPT Styler</h1>
-            {(page === "Text Editor" || page === "Message Editor") && (
+            {page === "Message Editor" && (
                 <button
                     className={`${css.btn} absolute inset-y-0 right-0 animate-fade-in m-2`}
                     onClick={() => setPage("")}
