+++ src/popup/views/messageEditor/components/MessageSliderControls.tsx
