+++ src/popup/views/messageEditor/styles.module.css
